+++ Tests/SentryTests/SentryCrash/SentryCrashInstallationReporterTests.swift
@@ -11,7 +11,7 @@ class SentryCrashInstallationReporterTests: XCTestCase {
     override func setUp() {
         super.setUp()
         sut = SentryCrashInstallationReporter(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: []), crashWrapper: TestSentryCrashWrapper.sharedInstance(), dispatchQueue: TestSentryDispatchQueueWrapper())
-        sut.install()
+        sut.install("")
         // Works only if SentryCrash is installed
         sentrycrash_deleteAllReports()
     }